+++ lib/util/mixins/movement_by_joystick.dart
@@ -8,7 +8,7 @@ mixin MovementByJoystick on Movement {
   static const REDUCTION_SPEED_DIAGONAL = 0.7;
 
   /// flag to set if you only want the 8 directions movement. Set to false to have full 360 movement
-  bool dPadAngles = true;
+  bool dPadAngles = false;
 
   /// the angle the player should move in 360 mode
   double movementRadAngle = 0;
@@ -38,9 +38,6 @@ mixin MovementByJoystick on Movement {
           if (innerCurrentDirectional != JoystickMoveDirectional.IDLE) {
             moveFromAngle(speed, movementRadAngle);
           }
-          // movement was done on the above line, this is only for the animation
-          // which is why we use zero speed as we don't want to translate position twice
-          _moveDirectional(innerCurrentDirectional!, 0, 0);
         }
       }
     }